requirements.txt
Django>=1.8.2
requests>=2.7.0
pytz==2015.4
mf2py==1.1.2
python-dateutil==2.8.0
BeautifulSoup==3.2.1
